refactor: Replace query_history with task_history (#2733)
* refactor: Replace query_history with task_history

* refactor: Remove dead code

* chore: Remove federation change because its not relevant

* refactor: Remove .replace for inputs because it's not actually a problem

* Apply suggestions from code review


* fix: Allow integration tests to run on feature branches

* explicit results_cache_hit label

---------
