+++ .github/workflows/integration.yml
@@ -6,6 +6,7 @@ on:
     branches:
       - trunk
       - release-*
+      - feature-*
 
   workflow_dispatch:
